[Lab9] DecadeMadness 계산 그 때 그때 더블 => int로

+++ Lab9/src/academy/pocu/comp2500/lab9/DecadeMadness.java
@@ -9,7 +9,7 @@ public class DecadeMadness extends PricingPolicy {
         if (books.isEmpty()) {
             return 0;
         }
-        double price = 0;
+        int price = 0;
         HashMap<Integer, ArrayList<Integer>> generationBooksMap = new HashMap<>();
         for (Book book : books) {
             int generation = book.getPublishedYear() / 10 * 10;
@@ -21,8 +21,8 @@ public class DecadeMadness extends PricingPolicy {
 
         for (ArrayList<Integer> prices : generationBooksMap.values()) {
             double priceToAdd = (prices.size() == 1) ? prices.get(0) : prices.stream().mapToDouble(p -> p).sum() * 0.8;
-            price += priceToAdd;
+            price += (int)priceToAdd;
         }
-        return (int) price;
+        return price;
     }
 }
